Initial Prototype
This is my initial attempt to build a quick and simply script to process the data files we generate for ClinicalTrials.gov. Starting out making this a thing just for me but with some work, this could be made more generalisable.

This rudimentary first version works with turning a single zipped CSV file into a processed CSV. Currently testing multiple files.