Switched to CSV module and multiple files
Got rid of extraneous packages and moved file creation over to only using the CSV module, removing the need for Pandas.

I've also tested to see if the code works if you have two files to process. With a bit of garbage collection so the console doesn't crash, this does work. Only tested within Jupyterlab in a Docker environment so far.